refactor: move get_device_name and get_interface_name
Above methods have been moved under utils sinece
they do not strictly belong to the firewall.
